Start updating bq load process so we can use it for 4337 data
