+++ src/op_analytics/datapipeline/etl/loadbq/main.py
@@ -2,7 +2,8 @@ from enum import Enum
 
 from op_analytics.coreutils.partitioned.location import DataLocation
 
-from .superchain_raw import load_superchain_raw_to_bq
+from op_analytics.datapipeline.etl.ingestion.reader.rootpaths import RootPath
+from .load import load_blockbatch_to_bq
 
 
 class PipelineStage(str, Enum):
@@ -30,3 +31,26 @@ def load_to_bq(
         )
 
     raise NotImplementedError()
+
+
+def load_superchain_raw_to_bq(
+    location: DataLocation,
+    range_spec: str,
+    dryrun: bool,
+    force_complete: bool,
+    force_not_ready: bool,
+):
+    return load_blockbatch_to_bq(
+        location=location,
+        range_spec=range_spec,
+        root_paths_to_read=[
+            RootPath.of("ingestion/blocks_v1"),
+            RootPath.of("ingestion/logs_v1"),
+            RootPath.of("ingestion/traces_v1"),
+            RootPath.of("ingestion/transactions_v1"),
+        ],
+        bq_dataset_name="superchain_raw",
+        dryrun=dryrun,
+        force_complete=force_complete,
+        force_not_ready=force_not_ready,
+    )
